+++ src/main/java/com/debs/repository/CustomerRepository.java
@@ -6,8 +6,12 @@ import org.springframework.stereotype.Repository;
 import com.debs.model.Customer;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+//	@Query(value = "select `name` from `customer`;", nativeQuery = true)
+//	List<String> findAllCustomersNames();
 
+//	@Query(value = "CALL AddCustomerAccount (:uid);", nativeQuery = true)
+//	int addCustomerAccount(@Param("uid") int uid);
 }
 
 //1: add method in service and serviceImpl
