+++ src/main/java/com/debs/repository/PartnerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.debs.model.Partner;
 
 @Repository
-public interface PartnerRepository extends JpaRepository<Partner, Long> {
+public interface PartnerRepository extends JpaRepository<Partner, Integer> {
 
 }
